feature[ats_ATS19845]: Convert Lib.generateUuid to generate nextInt

+++ plugins/org.eclipse.osee.framework.jdk.core/src/org/eclipse/osee/framework/jdk/core/util/Lib.java
@@ -1695,7 +1695,7 @@ public final class Lib {
    public static long generateUuid() {
       long next = 0;
       while (next <= 0) {
-         next = new Random().nextLong();
+         next = new Random().nextInt();
       }
       return next;
    }
